basic setup completed - routes, auth, users, DB, etc

+++ src/v1/services/users/users.controllers.js
@@ -1,13 +1,21 @@
+const boom = require('@hapi/boom')
 const UserService = require('./users.services')
+const { encryptPass } = require('../../../utils/auth.utils')
 
 const service = new UserService()
 
 class UserController {
     async create(req, res, next) {
+        const { body } = req
         try {
-            const body = req.body
+            const userExist = await service.findByEmail(body.email)
+            if(userExist) throw boom.conflict('user already exist')
+
+            body.password = encryptPass(body.password)
             const newUser = await service.create(body)
-            res.status(201).json(newUser)
+            const {password, ...userWithoutPassword} = newUser.dataValues
+
+            res.status(201).json(userWithoutPassword)
         } catch(error) {
             next(error)
         }
@@ -16,7 +24,7 @@ class UserController {
     async findAll(req, res, next) {
         try {
             const users = await service.findAll()
-            res.json(users)
+            res.status(200).json(users)
         } catch(error) {
             next(error)
         }
@@ -26,7 +34,10 @@ class UserController {
         try {
             const { id } = req.params
             const user = await service.findOne(id)
-            res.json(user)
+            if(!user) throw boom.notFound('user not found')
+
+            const {password, ...userWithoutPassword} = user.dataValues
+            res.status(200).json(userWithoutPassword)
         } catch(error) {
             next(error)
         }
@@ -35,9 +46,12 @@ class UserController {
     async updateOne(req, res, next) {
         try {
             const { id } = req.params
-            const body = req.body
-            const user = await service.updateOne(id, body)
-            res.json(user)
+            const userExist = await service.findOne(id)
+            if(!userExist) throw boom.notFound('user not found')
+
+            const user = await service.updateOne(id, req.body)
+            const {password, ...userWithoutPassword} = user.dataValues
+            res.status(201).json(userWithoutPassword)
         } catch(error) {
             next(error)
         }
@@ -46,8 +60,11 @@ class UserController {
     async deleteOne(req, res, next) {
         try {
             const { id } = req.params
+            const userExist = await service.findOne(id)
+            if(!userExist) throw boom.notFound('user not found')
+
             await service.deleteOne(id)
-            res.status(201).json({ id })
+            res.status(204).json({ id, message: 'user deleted' })
         } catch(error) {
             next(error)
         }
